+++ apps/c/CloverLeaf_3D_HDF5/MPI_inline/advec_cell_kernel1_zdir_mpiinline_kernel.cpp
@@ -126,64 +126,100 @@ void ops_par_loop_advec_cell_kernel1_zdir(char const *name, ops_block block, int
 
 
   //set up initial pointers and exchange halos if necessary
-  int base0 = args[0].dat->base_offset + (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) * start[0] * args[0].stencil->stride[0];
-  base0 = base0+ (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) *
-    args[0].dat->size[0] *
-    start[1] * args[0].stencil->stride[1];
-  base0 = base0+ (block->instance->OPS_soa ? args[0].dat->type_size : args[0].dat->elem_size) *
-    args[0].dat->size[0] *
-    args[0].dat->size[1] *
-    start[2] * args[0].stencil->stride[2];
+  long long int base0 =
+      args[0].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[0].dat->type_size
+                                                     : args[0].dat->elem_size) *
+                                     start[0] * args[0].stencil->stride[0];
+  base0 = base0 +
+          (long long int)(block->instance->OPS_soa ? args[0].dat->type_size
+                                                   : args[0].dat->elem_size) *
+              args[0].dat->size[0] * start[1] * args[0].stencil->stride[1];
+  base0 = base0 + (long long int)(block->instance->OPS_soa
+                                      ? args[0].dat->type_size
+                                      : args[0].dat->elem_size) *
+                      args[0].dat->size[0] * args[0].dat->size[1] * start[2] *
+                      args[0].stencil->stride[2];
   double *p_a0 = (double *)(args[0].data + base0);
 
-  int base1 = args[1].dat->base_offset + (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) * start[0] * args[1].stencil->stride[0];
-  base1 = base1+ (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) *
-    args[1].dat->size[0] *
-    start[1] * args[1].stencil->stride[1];
-  base1 = base1+ (block->instance->OPS_soa ? args[1].dat->type_size : args[1].dat->elem_size) *
-    args[1].dat->size[0] *
-    args[1].dat->size[1] *
-    start[2] * args[1].stencil->stride[2];
+  long long int base1 =
+      args[1].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[1].dat->type_size
+                                                     : args[1].dat->elem_size) *
+                                     start[0] * args[1].stencil->stride[0];
+  base1 = base1 +
+          (long long int)(block->instance->OPS_soa ? args[1].dat->type_size
+                                                   : args[1].dat->elem_size) *
+              args[1].dat->size[0] * start[1] * args[1].stencil->stride[1];
+  base1 = base1 + (long long int)(block->instance->OPS_soa
+                                      ? args[1].dat->type_size
+                                      : args[1].dat->elem_size) *
+                      args[1].dat->size[0] * args[1].dat->size[1] * start[2] *
+                      args[1].stencil->stride[2];
   double *p_a1 = (double *)(args[1].data + base1);
 
-  int base2 = args[2].dat->base_offset + (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size) * start[0] * args[2].stencil->stride[0];
-  base2 = base2+ (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size) *
-    args[2].dat->size[0] *
-    start[1] * args[2].stencil->stride[1];
-  base2 = base2+ (block->instance->OPS_soa ? args[2].dat->type_size : args[2].dat->elem_size) *
-    args[2].dat->size[0] *
-    args[2].dat->size[1] *
-    start[2] * args[2].stencil->stride[2];
+  long long int base2 =
+      args[2].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[2].dat->type_size
+                                                     : args[2].dat->elem_size) *
+                                     start[0] * args[2].stencil->stride[0];
+  base2 = base2 +
+          (long long int)(block->instance->OPS_soa ? args[2].dat->type_size
+                                                   : args[2].dat->elem_size) *
+              args[2].dat->size[0] * start[1] * args[2].stencil->stride[1];
+  base2 = base2 + (long long int)(block->instance->OPS_soa
+                                      ? args[2].dat->type_size
+                                      : args[2].dat->elem_size) *
+                      args[2].dat->size[0] * args[2].dat->size[1] * start[2] *
+                      args[2].stencil->stride[2];
   double *p_a2 = (double *)(args[2].data + base2);
 
-  int base3 = args[3].dat->base_offset + (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size) * start[0] * args[3].stencil->stride[0];
-  base3 = base3+ (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size) *
-    args[3].dat->size[0] *
-    start[1] * args[3].stencil->stride[1];
-  base3 = base3+ (block->instance->OPS_soa ? args[3].dat->type_size : args[3].dat->elem_size) *
-    args[3].dat->size[0] *
-    args[3].dat->size[1] *
-    start[2] * args[3].stencil->stride[2];
+  long long int base3 =
+      args[3].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[3].dat->type_size
+                                                     : args[3].dat->elem_size) *
+                                     start[0] * args[3].stencil->stride[0];
+  base3 = base3 +
+          (long long int)(block->instance->OPS_soa ? args[3].dat->type_size
+                                                   : args[3].dat->elem_size) *
+              args[3].dat->size[0] * start[1] * args[3].stencil->stride[1];
+  base3 = base3 + (long long int)(block->instance->OPS_soa
+                                      ? args[3].dat->type_size
+                                      : args[3].dat->elem_size) *
+                      args[3].dat->size[0] * args[3].dat->size[1] * start[2] *
+                      args[3].stencil->stride[2];
   double *p_a3 = (double *)(args[3].data + base3);
 
-  int base4 = args[4].dat->base_offset + (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) * start[0] * args[4].stencil->stride[0];
-  base4 = base4+ (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) *
-    args[4].dat->size[0] *
-    start[1] * args[4].stencil->stride[1];
-  base4 = base4+ (block->instance->OPS_soa ? args[4].dat->type_size : args[4].dat->elem_size) *
-    args[4].dat->size[0] *
-    args[4].dat->size[1] *
-    start[2] * args[4].stencil->stride[2];
+  long long int base4 =
+      args[4].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[4].dat->type_size
+                                                     : args[4].dat->elem_size) *
+                                     start[0] * args[4].stencil->stride[0];
+  base4 = base4 +
+          (long long int)(block->instance->OPS_soa ? args[4].dat->type_size
+                                                   : args[4].dat->elem_size) *
+              args[4].dat->size[0] * start[1] * args[4].stencil->stride[1];
+  base4 = base4 + (long long int)(block->instance->OPS_soa
+                                      ? args[4].dat->type_size
+                                      : args[4].dat->elem_size) *
+                      args[4].dat->size[0] * args[4].dat->size[1] * start[2] *
+                      args[4].stencil->stride[2];
   double *p_a4 = (double *)(args[4].data + base4);
 
-  int base5 = args[5].dat->base_offset + (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size) * start[0] * args[5].stencil->stride[0];
-  base5 = base5+ (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size) *
-    args[5].dat->size[0] *
-    start[1] * args[5].stencil->stride[1];
-  base5 = base5+ (block->instance->OPS_soa ? args[5].dat->type_size : args[5].dat->elem_size) *
-    args[5].dat->size[0] *
-    args[5].dat->size[1] *
-    start[2] * args[5].stencil->stride[2];
+  long long int base5 =
+      args[5].dat->base_offset + (long long int)(block->instance->OPS_soa
+                                                     ? args[5].dat->type_size
+                                                     : args[5].dat->elem_size) *
+                                     start[0] * args[5].stencil->stride[0];
+  base5 = base5 +
+          (long long int)(block->instance->OPS_soa ? args[5].dat->type_size
+                                                   : args[5].dat->elem_size) *
+              args[5].dat->size[0] * start[1] * args[5].stencil->stride[1];
+  base5 = base5 + (long long int)(block->instance->OPS_soa
+                                      ? args[5].dat->type_size
+                                      : args[5].dat->elem_size) *
+                      args[5].dat->size[0] * args[5].dat->size[1] * start[2] *
+                      args[5].stencil->stride[2];
   double *p_a5 = (double *)(args[5].data + base5);
